added Body component containing function to return current page
